feat: Create pure component (Todo Item)
